It compiles :)
