+++ WeaponMechanics/src/main/java/me/deecaad/weaponmechanics/lib/MythicMobsWeaponShootSkill.java
@@ -5,6 +5,7 @@ import io.lumine.mythic.api.adapters.AbstractLocation;
 import io.lumine.mythic.api.config.MythicLineConfig;
 import io.lumine.mythic.api.skills.*;
 import io.lumine.mythic.bukkit.BukkitAdapter;
+import me.deecaad.core.compatibility.CompatibilityAPI;
 import me.deecaad.core.utils.NumberUtil;
 import me.deecaad.weaponmechanics.WeaponMechanics;
 import me.deecaad.weaponmechanics.WeaponMechanicsAPI;
@@ -48,7 +49,7 @@ public class MythicMobsWeaponShootSkill implements ITargetedEntitySkill, ITarget
             target = ((LivingEntity) entity).getEyeLocation();
         } else {
             target = entity.getLocation();
-            target.setY(target.getY() + WeaponCompatibilityAPI.getWeaponCompatibility().getHeight(entity) / 2.0);
+            target.setY(target.getY() + CompatibilityAPI.getEntityCompatibility().getHeight(entity) / 2.0);
         }
 
         return castAtLocation(skillMetadata, BukkitAdapter.adapt(target));
