+++ WeaponMechanics/src/main/java/me/deecaad/weaponmechanics/listeners/WeaponListeners.java
@@ -1,8 +1,9 @@
 package me.deecaad.weaponmechanics.listeners;
 
 import me.deecaad.core.events.EntityEquipmentEvent;
+import me.deecaad.core.mechanics.CastData;
+import me.deecaad.core.mechanics.Mechanics;
 import me.deecaad.weaponmechanics.WeaponMechanics;
-import me.deecaad.weaponmechanics.mechanics.Mechanics;
 import me.deecaad.weaponmechanics.utils.MetadataKey;
 import me.deecaad.weaponmechanics.weapon.WeaponHandler;
 import me.deecaad.weaponmechanics.weapon.damage.AssistData;
@@ -78,7 +79,7 @@ public class WeaponListeners implements Listener {
 
             Mechanics equipMechanics = getConfigurations().getObject(weaponTitle + ".Info.Weapon_Equip_Mechanics", Mechanics.class);
             if (equipMechanics != null) {
-                equipMechanics.use(new CastData(entityWrapper, weaponTitle, weaponStack));
+                equipMechanics.use(new CastData(entity, weaponTitle, weaponStack));
                 alreadyUsedEquipMechanics = true;
             }
 
@@ -97,7 +98,7 @@ public class WeaponListeners implements Listener {
             // Don't use holster mechanics is equip mechanics were already used
             if (!alreadyUsedEquipMechanics) {
                 Mechanics holsterMechanics = getConfigurations().getObject(dequippedWeapon + ".Info.Weapon_Holster_Mechanics", Mechanics.class);
-                if (holsterMechanics != null) holsterMechanics.use(new CastData(entityWrapper, dequippedWeapon, dequipped));
+                if (holsterMechanics != null) holsterMechanics.use(new CastData(entity, dequippedWeapon, dequipped));
             }
         }
     }
